requirements.txt
langchain
langchain-community
requests
langchain-ollama
ddgs
openai-whisper
sounddevice
scipy
numpy